OOP Refactoring and GenTree Fixes
