+++ src/features/dataVisualizer/dataVisualizerUtils.ts
@@ -1,4 +1,5 @@
 import { Config } from './Config';
+import DataVisualizer from './dataVisualizer';
 import { Data, EmptyList, List, Pair } from './dataVisualizerTypes';
 import { is_list, is_pair } from './list';
 
@@ -28,9 +29,13 @@ export function isArray(data: Data): data is Array<Data> {
 }
 
 export function isFunction(data: Data): data is Function {
-  return typeof data === 'function';
+  if (typeof data === 'function') {
+    DataVisualizer.isGenTree = false;
+    DataVisualizer.isBinTree = false;
+    return true;
+  }
+  return false;
 }
-
 export function isPair(data: Data): data is Pair {
   return is_pair(data);
 }
