+++ src/features/dataVisualizer/tree/ArrayTreeNode.tsx
@@ -9,17 +9,7 @@ import { DrawableTreeNode } from './DrawableTreeNode';
  * Represents a node corresponding to a Source pair or array.
  */
 export class ArrayTreeNode extends DrawableTreeNode {
-  Colors: string[] = [
-    'black',
-    '#d81d1d',
-    '#e46510',
-    '#259530',
-    '#27d6e6',
-    '#0d54ed',
-    '#5a10d1',
-    '#e6148f',
-    '#870854'
-  ];
+  Colors: string[] = ['#d81d1d', '#e46510', '#25a232', '#0d54ed', '#e6148f', '#ad0ede'];
   createDrawable(
     x: number,
     y: number,
@@ -28,13 +18,7 @@ export class ArrayTreeNode extends DrawableTreeNode {
     colorIndex: number
   ): JSX.Element {
     let color = '';
-    // if (!partOfNode){
-    //   ArrayTreeNode.i++;
-    // }
-    // if (ArrayTreeNode.i>5){
-    //   ArrayTreeNode.i=0;
-    // }
-    color = this.Colors[colorIndex % this.Colors.length];
+    color = colorIndex === -1 ? 'black' : this.Colors[colorIndex % this.Colors.length];
     const arrayProps = { nodes: this.children ?? [], x, y, color };
     const arrayDrawable = <ArrayDrawable {...arrayProps}></ArrayDrawable>;
